Move placement events to store folder
